test: isolate SessionLifecycle OS contracts
